+++ src/main/resources/static/js/board/fileupload.js
