+++ src/main/resources/static/js/board/preview-editor.js
