+++ src/main/resources/static/js/board/preview-time.js
