+++ src/main/resources/static/js/chat/chat.js
