+++ src/main/resources/templates/attendance/fragments/attendanceTables.html
