+++ src/main/resources/templates/board/detail.html
