+++ src/main/resources/templates/board/fragments/boardSideMenu.html
