+++ src/main/resources/templates/board/list.html
