+++ src/main/resources/templates/chat/fragments/chatSideMenu.html
