+++ src/main/resources/templates/chat/list.html
@@ -45,7 +45,7 @@
 
 <th:block layout:fragment="javascript">
     <!-- 채팅.js -->
-    <script src="/js/yhj/chat/chat.js?v=<%=System.currentTimeMillis() %"></script>
+    <script src="/js/chat/chat.js?v=<%=System.currentTimeMillis() %"></script>
 </th:block>
 
 <th:block layout:fragment="content">
