+++ src/main/resources/templates/common/javascript.html
@@ -16,4 +16,4 @@
 <!-- Flatpickr Korean Locale -->
 <script src="https://cdn.jsdelivr.net/npm/flatpickr/dist/l10n/ko.js"></script>
 <!-- sideMenu 시간 / 출근버튼 -->
-<script src="/js/yhj/commute.js?v=<%=System.currentTimeMillis() %>"></script>
+<script src="/js/attendance/commute.js?v=<%=System.currentTimeMillis() %>"></script>
