+++ src/main/resources/templates/fragment/sideMenu.html
@@ -12,7 +12,7 @@
             <div class="sb-sidenav-menu">
 
             <div class="nav">
-                <th:block th:replace="~{/yhj/fragments/commuteTime::commuteTime}"></th:block>
+                <th:block th:replace="~{/yhj/fragments/attendanceTime::commuteTime}"></th:block>
 
                 <!--/* 그룹 소 제목 */-->
                 <div class="sb-sidenav-menu-heading">Core</div>
@@ -87,7 +87,7 @@
 
                 <th:block th:replace="~{calendar/fragments/kahSideMenu :: kahSideMenu}"></th:block>
 
-                <th:block th:replace="~{yhj/fragments/commuteSideMenu :: commuteSideMenu}"></th:block>
+                <th:block th:replace="~{yhj/fragments/attendanceSideMenu :: commuteSideMenu}"></th:block>
 
                 <th:block th:replace="~{cis/fragments/cisSideMenu :: cisSideMenu}"></th:block>
